Collapses multiple .scala files for settings
