--- src/main/scala/org/deepdive/settings/models/CalibrationSettings.scala
@@ -1,4 +1,0 @@
-package org.deepdive.settings
-
-/* Calibration Settings */
-case class CalibrationSettings(holdoutFraction: Double, holdoutQuery: Option[String], observationQuery: Option[String])
