--- src/main/scala/org/deepdive/settings/models/Connection.scala
@@ -1,4 +1,0 @@
-package org.deepdive.settings
-
-/* Database connection specifie in the settings */
-case class Connection(url: String, user: String, password: String)
