--- src/main/scala/org/deepdive/settings/models/ExtractionSettings.scala
@@ -1,4 +1,0 @@
-package org.deepdive.settings
-
-/* Extraction Settings */
-case class ExtractionSettings(extractors: List[Extractor], parallelism: Int)
