--- src/main/scala/org/deepdive/settings/models/Extractor.scala
@@ -1,21 +1,0 @@
-package org.deepdive.settings
-
-/* Extractor specified in the settings */
-case class Extractor(
-  name: String,
-  style: String,
-  outputRelation: String,
-  inputQuery: InputQuery,
-  udfDir: String,
-  udf: String,
-  parallelism: Int,
-  inputBatchSize: Int,
-  outputBatchSize: Int,
-  dependencies: Set[String],
-  beforeScript: Option[String] = None,
-  afterScript: Option[String] = None,
-  sqlQuery: String = "",
-  cmd: Option[String] = None,
-  loader: String = "",
-  loaderConfig: LoaderConfig = null
-  )
