--- src/main/scala/org/deepdive/settings/models/InferenceSettings.scala
@@ -1,4 +1,0 @@
-package org.deepdive.settings
-
-case class InferenceSettings(factors: List[FactorDesc], insertBatchSize: Option[Int],
-  skipLearning: Boolean = false, weightTable: String = "", parallelGrounding: Boolean = false)
