--- src/main/scala/org/deepdive/settings/models/InputQuery.scala
@@ -1,14 +1,0 @@
-package org.deepdive.settings
-
-import scala.language.implicitConversions
-
-sealed trait InputQuery
-
-case class CSVInputQuery(filename: String, seperator: Char) extends InputQuery
-case class DatastoreInputQuery(query: String) extends InputQuery
-
-object InputQuery {
-  implicit def stringToInputquery(str: String) : InputQuery = {
-    InputQueryParser.parse(InputQueryParser.inputQueryExpr, str).get
-  }
-}
