--- src/main/scala/org/deepdive/settings/models/LoaderConfig.scala
@@ -1,9 +1,0 @@
-package org.deepdive.settings
-
-/* Extractor specified in the settings */
-case class LoaderConfig (
-    connection: String,
-    schemaFile: String,
-    threads: Int,
-    parallelTransactions: Int
-)
