--- src/main/scala/org/deepdive/settings/models/PipelineSettings.scala
@@ -1,12 +1,0 @@
-package org.deepdive.settings
-
-/* A DeepDive pipeline that specifies which extractors and factor tasks should be executed */
-case class Pipeline(id: String, tasks: Set[String])
-
-/* User settings pipelines */
-case class PipelineSettings(activePipelineName: Option[String], pipelines: List[Pipeline],
-  relearnFrom: String, baseDir: Option[String]) {
-  def activePipeline : Option[Pipeline] = activePipelineName.flatMap { name =>
-    pipelines.find(_.id == name)
-  }
-}
