--- src/main/scala/org/deepdive/settings/models/SamplerSettings.scala
@@ -1,4 +1,0 @@
-package org.deepdive.settings
-
-/* Calibration Settings */
-case class SamplerSettings(samplerCmd: String, samplerArgs: String)
