--- src/main/scala/org/deepdive/settings/parsing/DataTypeParser.scala
@@ -1,10 +1,0 @@
-package org.deepdive.settings
-
-import scala.util.parsing.combinator.RegexParsers
-import org.deepdive.Context
-
-object DataTypeParser extends RegexParsers {
-  def CategoricalParser = "Categorical" ~> "(" ~> """\d+""".r <~ ")" ^^ { n => MultinomialType(n.toInt) }
-  def BooleanParser = "Boolean" ^^ { s => BooleanType }
-  def dataType = CategoricalParser | BooleanParser
-}
